+++ explorer/src/lib/components/__tests__/TransactionsCard.spec.js
@@ -16,6 +16,7 @@ describe("Transactions Card", () => {
 
   const baseProps = {
     error: null,
+    isSmallScreen: false,
     loading: false,
     txns: null,
   };