+++ explorer/src/lib/components/blocks-card/BlocksCard.svelte
@@ -14,6 +14,9 @@
   /** @type {Boolean} */
   export let loading;
 
+  /** @type {boolean} */
+  export let isSmallScreen;
+
   const ITEMS_TO_DISPLAY = 15;
 
   let itemsToDisplay = ITEMS_TO_DISPLAY;
@@ -44,14 +47,13 @@
     label: "Show More",
   }}
 >
-  <BlocksTable
-    data={displayedBlocks}
-    className="blocks-card__table mobile-hidden"
-  />
-
-  <div class="blocks-card__list desktop-hidden">
-    {#each displayedBlocks as block (block)}
-      <BlocksList data={block} />
-    {/each}
-  </div>
+  {#if isSmallScreen}
+    <div class="blocks-card__list">
+      {#each displayedBlocks as block (block)}
+        <BlocksList data={block} />
+      {/each}
+    </div>
+  {:else}
+    <BlocksTable data={displayedBlocks} className="blocks-card__table" />
+  {/if}
 </DataCard>