+++ explorer/src/routes/blocks/block/+page.svelte
@@ -30,6 +30,7 @@
     $navigating.complete.then(updateData);
   }
 
+  $: ({ isSmallScreen } = $appStore);
   $: ({ data, error, isLoading } = $dataStore);
 </script>
 
@@ -45,6 +46,7 @@
         {error}
         loading={isLoading}
         isOnHomeScreen={false}
+        {isSmallScreen}
         displayTooltips={true}
       />
     </Rerender>